Refactoring (arrangement of test classes, deletion of GroupList, Generalization of so methods, outsourcuing of registration actions from Program.cs,...) #10
