+++ ChildrenManagement/data/files/children.json
@@ -2,13 +2,11 @@
   {
     "ID": 1234567894561,
     "Name": "Geoffroy",
-    "Firstname": "Amélia",
+    "Firstname": "Appolline",
     "Nationality": 0,
-    "BirthDate": "2023-06-23T00:00:00",
+    "BirthDate": "2023-07-23T00:00:00",
     "PicturePath": "",
     "GroupName": "Les Nooooooon!",
-    "TrustedPeopleIds": [
-      7894561234562
-    ]
+    "TrustedPeopleIds": [7894561234562]
   }
 ]
