+++ ChildrenManagement/data/files/trustedPeople.json
@@ -2,10 +2,10 @@
   {
     "ID": 7894561234562,
     "Name": "Geoffroy",
-    "Firstname": "Elise",
+    "Firstname": "Madeleine",
     "Nationality": 0,
     "RelationshipToChild": 0,
-    "PhoneNumber": "+33672388440",
+    "PhoneNumber": "+33673366552",
     "BirthDate": "1992-07-17T00:00:00"
   }
 ]
