+++ ChildrenManagementTest/MSTestSettings.cs
@@ -1,1 +1,1 @@
-﻿[assembly: Parallelize(Scope = ExecutionScope.ClassLevel)]
+﻿
